+++ Amman/Assets/tuzinokuni/New Folder/P_dame.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fbddd37b49b2a5144bd3121bfaf6af2e
+guid: 31834b03e811113448a0db6b0e763358
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
