+++ Amman/ProjectSettings/TagManager.asset
@@ -8,6 +8,7 @@ TagManager:
   - wall
   - heart
   - b
+  - Enemy
   layers:
   - Default
   - TransparentFX
@@ -17,8 +18,8 @@ TagManager:
   - UI
   - Player
   - wall
-  - 
-  - 
+  - Playerdame
+  - Enemy
   - 
   - 
   -
